requirements.txt
openai>=0.27.0
transformers>=4.30.0
torch>=1.13.0
typer[all]>=0.9.0
pydantic>=1.10.0
pytest>=7.0.0
python-dotenv>=1.0.0
perplexityai>=0.13.0

fastapi>=0.95.0
uvicorn[standard]>=0.22.0
